+++ crates/nose-normalize/src/value_graph.rs
@@ -79,8 +79,9 @@ use nose_semantics::{
     admitted_js_like_map_constructor_at_call, admitted_js_like_set_constructor_at_call,
     admitted_library_method_call_at_call, admitted_map_get_at_call_span,
     admitted_map_key_view_at_call_span, admitted_map_key_view_wrapper_at_call_span,
-    admitted_ruby_set_factory_at_call_span, admitted_rust_option_and_then_at_call,
-    admitted_rust_option_none_sentinel_at_node, admitted_rust_option_some_constructor_at_call,
+    admitted_property_builtin_at_field, admitted_ruby_set_factory_at_call_span,
+    admitted_rust_option_and_then_at_call, admitted_rust_option_none_sentinel_at_node,
+    admitted_rust_option_some_constructor_at_call, admitted_rust_option_some_constructor_at_node,
     admitted_rust_vec_macro_factory_at_call_span, admitted_rust_vec_new_factory_at_call,
     admitted_scalar_integer_method_at_call, admitted_static_index_membership_at_call,
     asserted_unshadowed_global_symbol, binding_write_target, builder_append_call_args, builtin_tag,
@@ -90,17 +91,15 @@ use nose_semantics::{
     go_zero_map_default_kind, go_zero_map_entry_contract_for_node,
     go_zero_map_literal_contract_for_node, go_zero_map_lookup_contract, import_fact_evidence_rhs,
     imported_literal_producer_evidence_for_node, imported_namespace_symbol,
-    library_api_contract_evidence_for_node, library_property_builtin_contract,
-    library_rust_option_some_constructor_contract, map_builder_index_write_contract,
-    nullish_global_contract, opaque_argument_escape_args, own_property_guard_for_node,
-    receiver_mutation_call_receiver, record_shape_guard_for_node, reduction_builtin_contract,
-    semantics, seq_surface_contract_for_node, source_comprehension_at_node,
-    source_operator_at_node, source_pattern_at_node, source_range_at_node,
-    unproven_membership_like_method_contract, BuiltinArgContract, CBytePackWidth,
-    CardinalityPredicate, CardinalityThreshold, ComparisonLaw, DomainEvidence, DomainRequirement,
-    GoZeroMapDefaultKind, ImportFactKind, ImportedNamespaceFunctionSemantic,
+    map_builder_index_write_contract, nullish_global_contract, opaque_argument_escape_args,
+    own_property_guard_for_node, receiver_mutation_call_receiver, record_shape_guard_for_node,
+    reduction_builtin_contract, semantics, seq_surface_contract_for_node,
+    source_comprehension_at_node, source_operator_at_node, source_pattern_at_node,
+    source_range_at_node, unproven_membership_like_method_contract, BuiltinArgContract,
+    CBytePackWidth, CardinalityPredicate, CardinalityThreshold, ComparisonLaw, DomainEvidence,
+    DomainRequirement, GoZeroMapDefaultKind, ImportFactKind, ImportedNamespaceFunctionSemantic,
     IndexMembershipThreshold, IndexWriteReceiverContract, IteratorAdapterReceiverContract,
-    JavaMapFactoryKind, LibraryApiCalleeContract, LibraryApiEvidenceStatus, LibraryApiSpanCall,
+    JavaMapFactoryKind, LibraryApiCalleeContract, LibraryApiSpanCall,
     LibraryCollectionFactoryResult, LibraryMapFactoryResult, MapKeyViewKind, MethodBuiltinArgs,
     MethodReceiverContract, MethodSemanticContract, ReductionBuiltinContract, ScalarIntegerMethod,
     SeqSurfaceContract, StaticIndexMembershipKind, ValueDomain, ValueLaw, SEQ_VALUE_COLLECTION,
